+++ .eslintrc.js
@@ -23,6 +23,7 @@ module.exports = {
   ],
   extends: [
     "eslint:recommended",
+    "next",
     "plugin:@typescript-eslint/eslint-recommended",
     "plugin:@typescript-eslint/recommended",
     "plugin:jest-dom/recommended",