+++ src/components/Layout/Layout.tsx
@@ -5,17 +5,15 @@ import {
 } from "@nypl/design-system-react-components";
 import Feedback from "~/src/components/Feedback/Feedback";
 import { useRouter } from "next/router";
-import { Header, navConfig } from "@nypl/dgx-header-component";
 
 /**
  * Container class providing header, footer,
  * and other set up information to all its children.
  */
 
-const Layout: React.FC<{ children; isTestMode?: boolean }> = ({
-  children,
-  isTestMode = false,
-}) => {
+const Layout: React.FC<{
+  children?: React.ReactNode;
+}> = ({ children }) => {
   const router = useRouter();
   const [loading, setLoading] = useState(true);
 
@@ -41,22 +39,13 @@ const Layout: React.FC<{ children; isTestMode?: boolean }> = ({
   return (
     <>
       <DSProvider>
-        {!isTestMode && (
-          <Header
-            urlType="absolute"
-            skipNav={{ target: "main-content" }}
-            navData={navConfig.current}
-          />
+        {router.isFallback || loading ? (
+          <>
+            <SkeletonLoader />
+          </>
+        ) : (
+          <>{children}</>
         )}
-        <main id="main-content">
-          {router.isFallback || loading ? (
-            <>
-              <SkeletonLoader />
-            </>
-          ) : (
-            <>{children}</>
-          )}
-        </main>
         {!loading && <Feedback location={router.asPath} />}
       </DSProvider>
     </>