+++ src/components/WorkDetailDefinitionList/WorkDetailDefinitionList.test.tsx
@@ -2,9 +2,8 @@ import React from "react";
 import WorkDetailDefinitionList from "./WorkDetailDefinitionList";
 import "@testing-library/jest-dom/extend-expect";
 import { screen, render } from "@testing-library/react";
-import { ApiWork, WorkResult } from "~/src/types/WorkQuery";
-
-const apiWork: WorkResult = require("../../__tests__/fixtures/work-detail.json");
+import { ApiWork } from "~/src/types/WorkQuery";
+import { workDetail as apiWork } from "../../__tests__/fixtures/WorkDetailFixture";
 
 describe("Work Detail table with all information", () => {
   beforeEach(() => {