+++ src/constants/analytics.ts
@@ -1,0 +1,1 @@
+export const SITE_SECTION = "Digital Research Books";