+++ src/constants/links.ts
@@ -1,0 +1,2 @@
+export const SCAN_AND_DELIVER_LINK =
+  "https://www.nypl.org/research/scan-and-deliver";