+++ CHANGELOG.md
@@ -1,6 +1,40 @@
 # CHANGE LOG
 
-## [Pre-release]
+## [0.17.3]
+
+- Upgrade to Next 13.4.7
+- Add: SFR-1729: Verify the e-reader settings features for DRB
+- Replace the Feedback component with DS FeedbackBox component
+- Remove Redux for feedback component
+
+## [0.17.2]
+
+- Update to Web Reader version 4.3.1
+- Add TOC for single resource PDFs
+- Refactored Playwright/Cucumber tests to use new dropdown action function
+- Refactored Playwright/Cucumber tests to use new text fill function
+- Hide Header and Footer components on /read pages using the Web Reader
+- Migrate from Google Analytics to Adobe Analytics
+- Refactored Playwright/Cucumber tests to use new visibility function
+
+## [0.17.1]
+
+- Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app
+- Refactored Playwright/Cucumber tests to use new click action function
+
+## [0.17.0]
+
+- Replace Search Examples with Recently Added Collections to landing page
+- Add primary hero to the landing page
+- Add Part of Collection to Edition/Work pages
+- Add Collection page
+- Add collection placeholder image
+- Update hero image to update based on screen size
+- Chore: fresh package-lock.json install and fix install errors on mac with M1 chip
+- Add Playwright/Cucumber tests to support new Recently Added Collections feature on landing page
+- Refactored Playwright/Cucumber tests to use new page visit actions function
+
+## [0.16.1]
 
 - add features and tests for government docs filter
 - Add github actions to automate tag and release